requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
python-docx
pymupdf
langchain
langchain-openai
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
toml
scikit-learn
numpy
sentence-transformers
